project/ref_requirements.txt
asgiref==3.6.0
attrs==22.1.0
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.16.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.5
cryptography==40.0.2
Django==4.1.6
django-cors-headers==3.13.0
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
iniconfig==1.1.1
isodate==0.6.1
jsonfield==3.1.0
msal==1.21.0
msal-extensions==1.0.0
mysql-connector-python==8.0.32
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.3.0
pluggy==1.0.0
portalocker==2.7.0
psycopg2==2.9.5
py==1.11.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==306
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
